functions/battery_calculate_store/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
gobits==1.0.8
google-api-core==1.28.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-pubsub==2.3.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.0
idna==2.10
libcst==0.3.19
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.5
Werkzeug==1.0.1
